Explicit Close() when security-file-token provider writes tokens
This is a followup to a previous code review to clean up
the logic to explicitly Close() a file being written that
contains the authentication token to the secret store.
Deferred Close()s are removed and replaced with
explicit closes before leaving the function and an error-
reporting Close() at the end of each loop iteration.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>