Deprivilege token provider and revoke old tokens on startup. (#2269)
This commit modifies the back-half of the Vault init/unseal
flow so that the Vault key shares are used to generate a
transient root token when is then used for the remainder
of processing.  The transient root token is used to create
an admin token with the necessary privilege to run
security-file-token-provider with a non-root token.
Moreover, the transient token is also used for seeding
of database passwords and Kong TLS certificates.
Additionally, upon every startup, the existing admin/
service tokens from the previous run are revoked and
replaced with new ones.  The container entrypoint
script is changed to busy-wait at the end of its execution
so that tokens are not inadvertently revoked and replaced
should the container restart as happens during
blackbox testing.

This commit also includes hooks to revoke all root
tokens, including the one from the Vault initialization.
This hook is disabled by default as changes need to be
made to consuming microservices to consume their per-service
Vault tokens prior to enabling this feature.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>

Co-authored-by: tingyuz <37156729+tingyuz@users.noreply.github.com>