Implement Vault root token revokation
Changes revokeroottokens setting in docker configuration
from false to true and expunges the root token from
resp-init.json when revokeroottokens==true.

To facilitate this change, it is necessary to filter
the Vault /sys/init response so that the root_token
field can be removed.  Minor changes have been made
to SecretStoreClient such that all file IO is done
in secretstore-setup and the SecretStoreClient interfaces
data data structures instead of reader/writer interfaces.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>