Add In-Memory DB Implementation for Addressable
Fix #2286

Implement previously noop functionality for the memory.Client's methods
for dealing with persisting contract.Addressable types, and
memory.Client constructor.

Update memory.Client methods with noop implementation to panic rather
than return empty data. The reason is that the tests are using the newly
implemented constructor to create instances of the client for easier
maintainability. Previously, we were panicking at the constructor to
protect users from creating an instance of the memory.Client. Now we are
panicking at when an unimplemented method is invoked.

Add tests covering newly added implementation.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>