Update OpenAPI Docs for V2 API -- Ping, Metrics, Version, Configuration
Support for /ping, /metrics, /version, and /config endpoints is currently
specified for the HTTP GET method. That method is cachable. These are not
resource retrievals that should be cached; they are commands. Updated to
specify support via the HTTP POST method.

Added a corresponding request DTO for each of these endpoints.

Updated to support single and array of use-case requests and responses
for these endpoints.

Fixes: https://github.com/edgexfoundry/edgex-go/issues/2321

Signed-off-by: Michael Estrin <m.estrin@dell.com>