Update to latest go-mod-messaging
Fix #2339

Update to use the latest version of `go-mod-messaging` so we can
leverage the newly added functionality for MQTT support.

Update the existing `MessageQueueInfo` struct to include an `Optional`
field where additional configuration information can be passed into the
underlying client constructor. This new field is meant to facilitate
additional configurations that do not fit into the existing fields.

Update core data to pass the new `Optional` field to the client
constructor.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>