Remove InMemory Database Client Implementation
Fix #2340

Remove the InMemory implementation of the database client as we are
entertaining the idea of restructuring the database client for V2 of
edgex-go. Since InMemory implemenation was intended to be used for
acceptance testing, which will be available in V2, it is not needed now
and will need to change based on future design of the database
infrastructure layer.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>