Fix Redis Readings By Device & Value Descriptor
Fix #2344

Update the underlying query to leverage an existing function
`getObjectsByValuesSorted` which uses ZINTERSCORE and correctly returns
the desired data.

Fix issue with `getObjectsByValuesSorted` where the function would not
execute properly if the 'limit' was larger than the amount of 'ids'
found. This resulted in an index out of bounds issue.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>