snap: replace gosu with local drop-user cmd
The gosu commnand doesn't support libnss-extrausers
as used on Ubuntu Core, so the snap install fails
on an Ubuntu Core system when gosu is used to start
postgres.

This commit replaces gosu with a local C part which is
now used to run postgres and its applications as the
'snap_daemon' user.

Fixes: https://github.com/edgexfoundry/edgex-go/issues/2356

Signed-off-by: Tony Espy <espy@canonical.com>