snapcraft.yaml: fix builtin fuji device services
A recent change to device-sdk-go introduced new command-line
changes which break the device services (-random, and -virtual)
included by default in the snap. Both of these device services
were being built from the fuji branch instead of from a source
tag, which also contributed to this breakage. This commit updates
the command-line arguments for both, and also sets a git-tag to
be used for both instead of the fuji branch.

Fixes: https://github.com/edgexfoundry/edgex-go/issues/2366

Signed-off-by: Tony Espy <espy@canonical.com>