snap: update pre-refresh hook for geneva
The snap pre-fresh hook sets a snap config item
called 'release' to the current release name. If
we decide in the future to allow snap updates from
geneva to a future release, this variable would
allow the update code to determine which version
of the edgexfoundry snap was being updated.

Signed-off-by: Tony Espy <espy@canonical.com>