feat(security): Implement encryption of vault master key
This commit adds the ability to specify a hook that will be called
when intitializing the EdgeX secret store.  When enabled,
the resp-init.json contents will be encrypted by a set of
keys derived from a user-supplied input key that is
delivered over a process pipe.  The input key is expected to
be protected by some form of security hardware.

Fixes #1919

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>