Implement ability to add additional Kong routes
security-proxy-setup service now can take an environment variable, ADD_PROXY_ROUTE,
that takes a comma-separated list of paired additional service name and URL for which to create Kong routes.
The paired spec. is given like the following:

**\<ServiceName\>.\<RouteURL\>**

where ServiceName is the name of service which requests to create Kong route
and RouteURL is the full qualified URL for the service. eg.: http://edgex-app-service-rules:59807

So for a single service the value would be: "app-service-rules.http://edgex-app-service-rules:59807"

This service info is used to add the service to the routes for Kong.

Fixes: #2473

Address some PR feedback / comments

Signed-off-by: Jim Wang <yutsung.jim.wang@intel.com>