Deprecate Mongo Functionality (#2507)
Fix #2478

Add 'Deprecated' GoDoc comments to exported structs, functions, and
constants used exclusively for Mongo functionality. The comment signals
that the functionality has been deprecated as of Geneva. Methods have
not been included in the entities which have the updated GoDoc comments
since marking the struct the methods are tied to are marked as
deprecated. Support logging has not been modified as it still is
expected to use MongoDB for storing logging information.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>