Fix Reading Redis Binary Payload Persistence (#2551)
Fix #2527

Update the Redis DBClient to clear the `BinaryValue` of a reading before
marshaling the structure and storing it within a Key Value pair. This
approach does not affect the original Reading that is passed to the
DBClient as it is passed by value rather than by reference.

Add comments to Reading and Event structures, methods, and functions
which operate on the entities to mention the fact that Binary reading
data is not persisted and will be empty when performing GET operations.

Fix existing out-of-date Redis integration tests which were not
compiling nor passing(once compilation errors were fixed).

Update the flag used to enable the Redis integration tests to match the
Redis integration testing that was put into place in go-mod-messaging.

Update to the latest go-mod-core-contracts version which contains the
fix to the Reading construct to allow Readings to have empty
`BinaryValue` and still pass "Validation" when marshaling and
unmarshaling.

Update the Redis DBClient to not close the global channels during a
Disconnect. This was not an obvious issue since we are only using one
DBClient instance. However, if there are multiple instance then closing
one will cause the others to not function properly do to the use of a
closed channel. Also, other clients will fail to disconnect as closing
the already closed channel will result in a panic.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>