feat: Implement pluggable password generator
Remove dependency on GoKey for password generator;
if not specified, the built-in password generator
does the equivalent of "openssl rand -base64 33"
which generates a base64 encoded random string
of 264 random bits.

Otherwise, the code looks for a PasswordProvider
in configuration that points to an executable,
and PasswordProviderArgs that specify the
arguments to that executable.  The code then
launches that exectuable, trims the trailing
newline, and uses the result as a password.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>