feat(security): Remove Vault TLS requirement
Based on the ADR to remove TLS requirement for in-cluster communication, the Vault internally now will be listening on HTTP.
The protocol or scheme for security services and core services changes to http instead of https.
The related CA certificate and TLS cert key pair will be no longer needed and thus oboselete.

Closes: #2865
Signed-off-by: Jim Wang <yutsung.jim.wang@intel.com>