build(snap): update snap build pipeline patch
This change updates the snap build pipeline optimization
patch which when used in combination with the following
command:

$ snapcraft prime --use-lxd

...significantly reduces the snap build time when run
from the Jenkins pipeline. It accomplishes this by removing
the non-edgex-go parts from snapcraft.yaml. Additionally,
the 'snapcraft prime' command causes the snapcraft to prime
the snap, but stop before the actual binary snap package is
created.

Note - once this patch is actually put into place in
the pipeline builds, future updates to this patch will
accompany any future changes to snap/snapcraft.yaml.

Signed-off-by: Tony Espy <espy@canonical.com>