fix(snap): remove kong TLS config overrides
Due to the recent decision to remove use of TLS on a
single device/node, it was also decided to let Kong
generate its own self-signed certificate and get rid
of the EdgeX CA altogether. A PR landed this weekend
which updated some of the default configuration values
in security-secretstore-setup's configuration.toml file
related to TLS configuration. As the snap still contained
overrides for some of these files, installation of the
snap fail when security-secretstore-setup fails to start.
This change removes the configuration overrides related
to Kong TLS configuration from snapcraft.yaml, which then
allows the snap to be installed.

Signed-off-by: Tony Espy <espy@canonical.com>