fix(snap): add PR snap optimization script
This commit adds a new script which is meant for use by
the EdgeX CI pipeline builds to optimize the snap CI check
run for pull requests. It replaces the existing patch in
/snap/local/patches which does essentially the same thing,
but usually requires manual update when snapcraft.yaml
is modified.

The script performs the same gross edits as the match,
stripping out everything (apps and parts) from snapcraft.yaml
except what's needed to build/prime edgex-go, as the whole
idea of the CI check is to catch changes to edgex-go that
break the snap build.

Note - in addition to applying this patch, the pipeline also
only primes the snap (e.g. `snapcraft prime`), as there's no
need to build the finally binary .snap file (and it won't
work with the patch applied). This further reduces the build
time.

Signed-off-by: Tony Espy <espy@canonical.com>