feat(security): Secure Kong Admin API
Modified the secret-store process to create a declarative configuration file
for the Kong gateway configuration prior to start-up. The most logical place
for this configuration to take place was secretstore-setup, immediately after
vault was initialized.

To properly secure the Admin port, we first have to "hide" ports 8001/8444
by binding them to the Kong container. Doing that will only expose the ports
to the user that is running on the container --> kong. These changes will be
reflected in an adjoining PR on repo edgex-compose since they require
docker-compose changes.

Admin API access is needed by proxy-setup to create the services/routes to
the various microservices running in EdgeX. To enable that, a "loopback"
service/route combo was created that mapped "http://<host>:8000/admin" to
"http://<host>:8001". To avoid a chicken/egg problem, a declarative
configuration file is defined to import into Kong prior to starting the
Kong service for the first time. Values for an admin public key value and
JWT issuer value called out in the template are created and inserted.

The declarative configuration file enables plugins "JWT" and "ACL", and
creates a user (admin) and group (admin-group). All traffic to the "/admin"
path is locked down to the "admin-group" group and requires the use of a
Bearer token in the header of all requests to "/admin". A JWT is created
from an EC based public/private key pair, that is linked to the "admin"
user, and is written to the token path for proxy-setup.

Once the declarative configuration file is created with all necessary
values, it is written to "/tmp/kong/kong.yml" (changeable via config)
and is picked up by the Kong container through the use of the same
volume attachments between the 2 containers.

The Kong entrypoint script in bootstrap was modified to import the
declarative configuration file prior to "kong docker-start".

When Kong has been initialized, proxy-setup begins by reading in
the JWT file from its secure token directory, and then using that
JWT as the "Bearer" token via "Authorization" header in the API
requests to setup all of the gateway routes.

Closes: #2537

Signed-off-by: Beau Frusetta <beau.frusetta@intel.com>