feat(snap): add vault secret-token content interface
This content interface provides a mechanism for the edgexfoundry
snap to shared vault secret tokens in order for services in external
edgex snaps to access the secret-store. Note, in this case this snap
defines a plug instead of slot to allow the consuming snap to create
the service-specific directory under $SNAP_DATA/secrets.

Signed-off-by: Tony Espy <espy@canonical.com>