fix(snap): go 1.16 update (#3525)
* fix(snap): go 1.16 update

This change updates the snap to use go 1.16. Changes include:

  * updating build-snap to go 1.16
  * pulling consul from a prebuilt binary release file
  * building device-virtual from the master branch
  * set device-virtual app adapter to full (required to
    provide LD_LIBRARY_PATH and other env vars to app,
    needed to find libzmq)
  * using stage-snaps to include app-service-config

* feat(snap): add vault secret-token content interface

This content interface provides a mechanism for the edgexfoundry
snap to shared vault secret tokens in order for services in external
edgex snaps to access the secret-store. Note, in this case this snap
defines a plug instead of slot to allow the consuming snap to create
the service-specific directory under $SNAP_DATA/secrets.

Signed-off-by: Tony Espy <espy@canonical.com>

* fix(snap): set version from git tag

Signed-off-by: Tony Espy <espy@canonical.com>