fix(snap): create empty redis.conf file
This commit modifies the install hook to use a shell
script to create configure redis. This script creates
an empty redis.conf file in $SNAP_DATA/redis/conf.
This allows us a static command-line which always
specifies a config file path. If the secret-store is
enabled, this file will be just be overwritten.

Signed-off-by: Tony Espy <espy@canonical.com>