fix(snap): support disabling security
This commit adds support for disabling security
at runtime via the snap command:

$ sudo snap set edgexfoundry security-secret-store=off

Note - this is only meant to allow a developer to
disable security during development in order to reduce
complexity. Once disabled, the snap will not allow
security to be re-enabled (i.e. a re-install is the
only way to do so).

Fixes: #3712

Signed-off-by: Tony Espy <espy@canonical.com>