fix(security): update kong admin token file permissions (#3819)
* fix(security): make kong-admin-jwt writable

The kong-admin-jwt token is written using the
octal permission (0400) which makes it an
owner read-only file. The problem comes when
the file is written to persistent storage. If
EdgeX is running natively or via snap package,
security-secretstore-setup will fail when it
tries to write a new token.

Fixes: #3818

Signed-off-by: Tony Espy <espy@canonical.com>

* fix(snap): add post-refresh script

This script removes an existing kong-
admin-jwt file (which may be read-only),
allowing snap refreshes (and other
scenarios involving service restarts)
to function properly.

Signed-off-by: Tony Espy <espy@canonical.com>