feat!: Change identity source for secrets-config proxy adduser/deluser
BREAKING CHANGE: Remove secrets-config proxy jwt subcommand
and change the adduser/deluser commands so that they
create vault identities and logins in Vault
instead of creating them in Kong. This change will make it
impossible to create a JWT that works with Kong proxy,
and will break security-enabled TAF tests.

This is the third commit in a sequence of commits to
implement the microservice authentication (token-based) ADR.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>