fix: Remove broken unit test for mosquitto password file
The unit test as fundamentically broken, as it
performed a real os.Chdir() to attempt to change file system paths,
as well as performed a real exec.Command() to run
the mosquitto_passwd utility with various arguments.
In effect, the removed unit test doesn't test the business
logic of SetupMosquittoPasswordFile,
but instead tests the behavior of chdir and mosquitto_passwd,
which are system-provided functionality.

Closes #4467

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>