fix!: Disable the default discovery scheduler (#4860)
The default auto discovery is disabled, but the default discovery is
very short (30s).  When the user enables the auto discovery, it would
trigger it every 30s by default.  This value doesn't fit the most of
use cases.  By modifying it to 0s, the discovery scheduler will be
disabled by default.

BREAKING CHANGE: If the user develops a Device Service with the auto
discovery mechanism and enable it, the discovery won't be automatically
triggered every 30s anymore.  Users have to define this
Device_Discovery_Interval: 30s in the Device Service to keep the
original behavior.

Signed-off-by: Cloud Tsai <cloudxxx8@gmail.com>