feat: revise DB schema initialVersion to 4.0.0-dev
related to https://github.com/edgexfoundry/edgex-go/issues/5072

The orignal initialVersion is specified as 4.0.0, which will cause
services with version 4.0.0-dev.XX below the db schema version and fails
to initialize the postgreSQL client. This will cause severla services
fail to bootstrap.

Specify the initialVersion to 4.0.0-dev to avoid this.

Signed-off-by: Jude Hung <jude@iotechsys.com>